feat(mobile): make all PR-link helper feedback toasts
Empty-clipboard paste, pasted non-URL text, and typed invalid links all
surface error toasts now; the inline helper-message UI and its reserved
slot are gone from the entry screen, so the layout has no message row at
all. The clear-button selector and the toast copy move to pr-link-paste
(pr-link-helper-slot is deleted). Open is disabled only when the field
is empty.
