+++ apps/mobile/src/components/pr-review/pr-review-entry-screen.tsx
@@ -19,17 +19,14 @@ import { Text } from '@/components/ui/text';
 import { parseGitHubPrUrl } from '@/lib/github-pr-url';
 import { useThemeColors } from '@/lib/hooks/use-theme-colors';
 import { getPrReviewPath } from '@/lib/profile-agent-navigation';
-import {
-  PR_LINK_HELPER_CLIPBOARD_EMPTY_COPY,
-  PR_LINK_HELPER_INVALID_COPY,
-  type PrLinkHelperMessage,
-  selectPrLinkClearButtonVisible,
-  selectPrLinkHelperSlotState,
-} from '@/lib/pr-review/pr-link-helper-slot';
 import { consumePrLinkInputEcho, pushPrLinkInputEcho } from '@/lib/pr-review/pr-link-input-echo';
-import { decidePrLinkPaste } from '@/lib/pr-review/pr-link-paste';
+import {
+  decidePrLinkPaste,
+  PR_LINK_TOAST_CLIPBOARD_EMPTY_COPY,
+  PR_LINK_TOAST_INVALID_COPY,
+  selectPrLinkClearButtonVisible,
+} from '@/lib/pr-review/pr-link-paste';
 import { getRecentPrs, type RecentPr, upsertRecentPr } from '@/lib/pr-review/recent-prs';
-import { cn } from '@/lib/utils';
 
 const URL_PLACEHOLDER = 'https://github.com/owner/repo/pull/123';
 
@@ -38,17 +35,16 @@ export function PrReviewEntryScreen() {
   const colors = useThemeColors();
   // Uncontrolled iOS input — keep the raw text in a ref so the submit
   // handler reads the latest value without re-rendering on every
-  // keystroke. State is only for derived UI (whether there's any text,
-  // active helper message). The TextInput component ref is for focus()
-  // and setNativeProps on programmatic paste.
+  // keystroke. State is only for derived UI (whether there's any text).
+  // The TextInput component ref is for focus() and setNativeProps on
+  // programmatic paste.
   const inputRef = useRef<TextInput>(null);
   const inputValueRef = useRef<string>('');
   // FIFO of values written via setNativeProps. Matching onChangeText values
-  // are treated as programmatic echoes (do not clear helpers / clobber ref).
+  // are treated as programmatic echoes (do not clobber ref).
   // Order-agnostic membership so double-taps and delayed native echoes work.
   const pendingProgrammaticTextsRef = useRef<string[]>([]);
   const [hasInput, setHasInput] = useState(false);
-  const [helperMessage, setHelperMessage] = useState<PrLinkHelperMessage | null>(null);
   const [recent, setRecent] = useState<RecentPr[] | null>(null);
 
   useFocusEffect(
@@ -80,10 +76,9 @@ export function PrReviewEntryScreen() {
     const raw = inputValueRef.current;
     const parsed = parseGitHubPrUrl(raw.trim());
     if (!parsed) {
-      setHelperMessage('invalid');
+      toast.error(PR_LINK_TOAST_INVALID_COPY);
       return;
     }
-    setHelperMessage(null);
     // Title is backfilled on first successful load (S5).
     await upsertRecentPr({
       owner: parsed.owner,
@@ -99,18 +94,15 @@ export function PrReviewEntryScreen() {
     const clipboard = await Clipboard.getStringAsync();
     const decision = decidePrLinkPaste(clipboard);
     if (decision.kind === 'empty') {
-      // Clipboard-empty is a toast, not an inline helper (layout slot stays
-      // reserved for invalid only).
-      toast.error(PR_LINK_HELPER_CLIPBOARD_EMPTY_COPY);
+      toast.error(PR_LINK_TOAST_CLIPBOARD_EMPTY_COPY);
       return;
     }
     // Replace entire field (never append-at-cursor): native field + ref + hasInput.
     applyFieldText(decision.text);
     if (decision.kind === 'non-url-text') {
-      setHelperMessage('invalid');
+      toast.error(PR_LINK_TOAST_INVALID_COPY);
       return;
     }
-    setHelperMessage(null);
     await handleSubmit();
   };
 
@@ -126,15 +118,7 @@ export function PrReviewEntryScreen() {
     router.push(getPrReviewPath(entry.owner, entry.repo, entry.number));
   };
 
-  const slotState = selectPrLinkHelperSlotState({
-    message: helperMessage,
-  });
   const showClearButton = selectPrLinkClearButtonVisible({ hasInput });
-  const isInvalid = helperMessage === 'invalid';
-  // Stable always-mounted slot (D6): same Text line box every state — only
-  // string + color token change. NBSP keeps the line box at every font scale.
-  // Clipboard-empty is a toast; slot is reserved for invalid only.
-  const helperSlotCopy = slotState === 'invalid' ? PR_LINK_HELPER_INVALID_COPY : '\u00A0';
 
   let recentsBody: ReactNode = null;
   if (recent === null) {
@@ -223,15 +207,11 @@ export function PrReviewEntryScreen() {
                     if (decision.kind === 'echo') {
                       // Echo of setNativeProps: inputValueRef already holds the
                       // intentional value from applyFieldText — do not clobber it
-                      // with a delayed/stale echo, and do not clear helpers.
+                      // with a delayed/stale echo.
                       return;
                     }
                     inputValueRef.current = value;
                     setHasInput(value.length > 0);
-                    // Any real edit clears the invalid helper message.
-                    if (helperMessage !== null) {
-                      setHelperMessage(null);
-                    }
                   }}
                   // leading-[normal] so no lineHeight reaches the style: an explicit lineHeight
                   // makes iOS draw the placeholder lower than the typed text (see AGENTS.md).
@@ -278,21 +258,8 @@ export function PrReviewEntryScreen() {
                 <ClipboardIcon size={18} color={colors.mutedForeground} />
               </Pressable>
             </View>
-            <Text
-              numberOfLines={1}
-              className={cn(
-                'text-sm',
-                slotState === 'invalid' ? 'text-destructive' : 'text-muted-foreground'
-              )}
-              // NBSP placeholder is blank to screen readers — hide only that state.
-              accessible={slotState !== 'none'}
-              accessibilityElementsHidden={slotState === 'none'}
-              importantForAccessibility={slotState === 'none' ? 'no-hide-descendants' : 'yes'}
-            >
-              {helperSlotCopy}
-            </Text>
             <Button
-              disabled={!hasInput || isInvalid}
+              disabled={!hasInput}
               onPress={() => {
                 void handleSubmit();
               }}
