+++ apps/mobile/src/lib/pr-review/pr-link-paste.ts
@@ -8,7 +8,7 @@ type PrLinkPasteDecision =
 /**
  * Decide how a paste-button tap should treat clipboard contents.
  * Trims first; empty → no insertion; valid PR URL → replace + navigate;
- * anything else → replace + invalid helper.
+ * anything else → replace + invalid toast at the call site.
  */
 export function decidePrLinkPaste(clipboard: string | null | undefined): PrLinkPasteDecision {
   const text = (clipboard ?? '').trim();
@@ -20,3 +20,21 @@ export function decidePrLinkPaste(clipboard: string | null | undefined): PrLinkP
   }
   return { kind: 'non-url-text', text };
 }
+
+type PrLinkClearButtonInput = {
+  /** Whether the uncontrolled PR-link field currently has any text. */
+  readonly hasInput: boolean;
+};
+
+/**
+ * Whether the in-field clear control should render.
+ * Present only when the field has content; absent when empty.
+ */
+export function selectPrLinkClearButtonVisible(input: PrLinkClearButtonInput): boolean {
+  return input.hasInput;
+}
+
+/** Toast copy when paste finds an empty clipboard. */
+export const PR_LINK_TOAST_CLIPBOARD_EMPTY_COPY = 'Clipboard is empty';
+/** Toast copy when paste or Open gets a non-PR link. */
+export const PR_LINK_TOAST_INVALID_COPY = 'Not a GitHub pull request link';
